+++ openharmony/arkcompiler_runtime_core/static_core/runtime/mem/gc/bitmap.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2025 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -56,7 +56,7 @@ public:
     {
         size_t countMarkBits = 0;
         for (BitmapWordType word : bitmap_) {
-            countMarkBits += Popcount(word);
+            countMarkBits += static_cast<size_t>(Popcount(word));
         }
         return countMarkBits;
     }
