+++ openharmony/arkcompiler_runtime_core/static_core/runtime/osr.cpp
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2025 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -168,7 +168,7 @@ extern "C" void *PrepareOsrEntry(const Frame *iframe, uintptr_t bcOffset, const 
                 break;
             case VRegInfo::Location::REGISTER:
                 // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
-                regBuffer[vreg.GetValue()] = value;
+                regBuffer[vreg.GetValue()] = static_cast<uint64_t>(value);
                 break;
             case VRegInfo::Location::FP_REGISTER:
                 // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
