+++ openharmony/arkui_ace_engine/frameworks/base/resource/shared_image_manager.cpp
@@ -127,7 +127,7 @@ bool SharedImageManager::UpdateImageMap(const std::string& name, const SharedIma
     } else {
         sharedImageTotalSize_ += sharedImage.size();
         sharedImageMap_.emplace(name, sharedImage);
-        if (sharedImageMap_.size() > sharedImageCacheThreshold_) {
+        if (static_cast<int>(sharedImageMap_.size()) > sharedImageCacheThreshold_) {
             LOGW("will clear %{public}s cache, sharedImageMap_ size %{public}d max cache: %{public}d",
                 name.c_str(), static_cast<int>(sharedImageMap_.size()), sharedImageCacheThreshold_);
             isClear = true;
