+++ openharmony/arkui_ace_engine/frameworks/core/common/window.cpp
@@ -21,7 +21,7 @@ namespace OHOS::Ace {
 Window::Window(std::unique_ptr<PlatformWindow> platformWindow) : platformWindow_(std::move(platformWindow))
 {
     CHECK_NULL_VOID(platformWindow_);
-    auto&& callback = [this](uint64_t nanoTimestamp, uint32_t frameCount) { OnVsync(nanoTimestamp, frameCount); };
+    auto&& callback = [this](uint64_t nanoTimestamp, uint64_t frameCount) { OnVsync(nanoTimestamp, frameCount); };
     platformWindow_->RegisterVsyncCallback(callback);
     LOGI("Window Created success.");
 }
@@ -43,7 +43,7 @@ void Window::SetRootRenderNode(const RefPtr<RenderNode>& root)
     platformWindow_->SetRootRenderNode(root);
 }
 
-void Window::OnVsync(uint64_t nanoTimestamp, uint32_t frameCount)
+void Window::OnVsync(uint64_t nanoTimestamp, uint64_t frameCount)
 {
     isRequestVsync_ = false;
